fix the problem when PWM length is longer than sequence lengths (#35)
* intro max_merged_length parameter; removed warnings

* removed warnings

* stop merging if merged motif is longer than the longest seq

* stop merging if merged motif is longer than the longest seq

+++ src/iupac_pattern.h
@@ -53,7 +53,7 @@ class IUPACPattern {
 
   size_t get_pattern();
   std::string get_pattern_string();
-  int get_pattern_length();
+  size_t get_pattern_length();
 
   float getExpCountFraction(const size_t pseudo_expected_pattern_counts);
   float getLogPval();
